Fix API key error
