+++ src/Input.js
@@ -8,17 +8,14 @@ export default class InputScene extends Phaser.Scene {
 
   create() {
     this.add.image(150, 80, 'logo').setScale(0.2);
-    this.add.text(100, 120, 'You did a great job, Evader!!!', {
-      color: '#FFF', fontFamily: 'Arial', fontSize: '12px ', fontWeight: '900',
-    });
     const element = document.getElementById('form');
-    element.style.display = 'flex';
+    // element.style.display = 'flex';
     element.addEventListener('click', (event) => {
       if (event.target.name === 'submit') {
         const user = document.getElementById('user');
         if (user.value !== '') {
           element.style.display = 'none';
-          scoreBoard.name(user.value);
+          scoreBoard.names(user.value);
           scoreBoard.postScores();
           this.scene.start('Title');
         } else {
