+++ src/Scene/BattleScene.js
@@ -1,5 +1,4 @@
 import Phaser from 'phaser';
-import scoreBoard from '../API/Scores';
 
 let userName;
 let finalScore;
@@ -86,13 +85,8 @@ class BattleScene extends Phaser.Scene {
 
     this.physics.add.collider(this.gameitems, obstacles);
 
-    // eslint-disable-next-line no-alert
-    userName = prompt('Please enter your name', 'name');
-    scoreBoard.name(userName);
-    scoreBoard.postScores();
-
     finalScore = this.add.text(26, 50, 'finalScore', { fontSize: '20px', fill: '#9f1239' });
-    finalScore.visible = false;
+    // finalScore.visible = false;
   }
 
   update() {
@@ -133,7 +127,7 @@ class BattleScene extends Phaser.Scene {
     star.disableBody(true, true);
     this.score += 100;
     this.scoreText.setText(`Score: ${this.score}`);
-    finalScore.setText(`${userName}, your core is: ${this.score}`);
+    this.finalScore.setText(`${userName}, your core is: ${this.score}`);
   }
 
   endGame(player) {
