+++ src/Scene/PreloadScene.js
@@ -101,7 +101,7 @@ export default class PreloaderScene extends Phaser.Scene {
   ready() {
     this.readyCount += 1;
     if (this.readyCount === 2) {
-      this.scene.start('Title');
+      this.scene.start('Input');
     }
   }
 }
